% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{National Health and Nutrition Examination Survey 2011-2012}
\format{
A data frame with 9756 rows and 16 variables:
\describe{
\item{seqn}{Respondent sequence number}
\item{ridageyr}{Age in years at screening}
\item{riagendr}{Gender}
\item{dmdmartl}{Marital status}
\item{dmdeduc2}{Education level - Adults 20+}
\item{sdmvpsu}{Masked variance pseudo-PSU}
\item{sdmvstra}{Masked variance pseudo-stratum}
\item{wtint2yr}{Full sample 2 year interview weight}
\item{female}{Gender}
\item{hsq496}{How many days feel anxious}
\item{hsq571}{SP donated blood in the past 12 months}
\item{hsd010}{General health condition}
\item{pad630}{Minutes moderate-intensity work}
\item{pad675}{Minutes moderate recreational activities}
\item{paq665}{Moderate recreational activities}
\item{pad680}{Minutes sedentary activity}
}
}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/Search/DataPage.aspx?Component=Demographics&CycleBeginYear=2011}
}
\usage{
nhanes
}
\description{
A dataset used in the UCLA Statistical Consulting Survey Analysis in R guide
\url{https://stats.idre.ucla.edu/r/seminars/survey-data-analysis-with-r/}
}
\keyword{datasets}
