% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ologit}
\alias{ologit}
\title{Simulated data for ordinal logistic regression example.}
\format{
A data frame with 400 rows and 4 variables:
\describe{
\item{apply}{Likelihood of applying to graduate school}
\item{pared}{Indicator for whether at least 1 parent has a graduate degree}
\item{public}{Indicator for whether undergraduate institution is public or private}
\item{gpa}{Student's grade point average}
}
}
\source{
\url{https://stats.idre.ucla.edu/stat/data/ologit.dta}
}
\usage{
ologit
}
\description{
A dataset used in the UCLA Statistical Consulting Ordinal Logistic Regression Example
\url{https://stats.idre.ucla.edu/r/dae/ordinal-logistic-regression/}
}
\keyword{datasets}
