% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomcheck.R
\name{pomcheck}
\alias{pomcheck}
\title{Title}
\usage{
pomcheck(dat, f)
}
\arguments{
\item{dat}{data frame containing the variables in the formula}

\item{f}{an object of class formula}
}
\value{
None
}
\description{
Title
}
\examples{
pomcheck(iris, Species ~ Sepal.Length)
}
