% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.pomcheck}
\alias{plot.pomcheck}
\title{Graphical check for proportional odds assumption}
\usage{
\method{plot}{pomcheck}(x, legend.position = "none", ...)
}
\arguments{
\item{x}{a pomcheck object}

\item{legend.position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector)}

\item{\dots}{currently unused}
}
\value{
None
}
\description{
Generates the plots described in \url{https://stats.idre.ucla.edu/r/dae/ordinal-logistic-regression/}
for checking if the proportional odds assumption holds for a cumulative logit model.
}
\examples{
plot(pomcheck(Species ~ Sepal.Width, iris))
}
\seealso{
\code{\link{pomcheck}}
}
