% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomcheck.R
\name{pomcheck}
\alias{pomcheck}
\title{Graphical check for proportional odds assumption}
\usage{
pomcheck(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments}
}
\value{
None
}
\description{
Graphical check for proportional odds assumption
}
\examples{
pomcheck(Species ~ Sepal.Length, iris)
pomcheck(x="Species", y="Sepal.Length", iris)
}
